+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Changelog of Git Changelog.
 
 ## Next release
+<<<<<<< HEAD
 ### Other changes
 
 **Show found refs when not finding ref**
@@ -26,13 +27,15 @@ Changelog of Git Changelog.
 
 
 ## 1.49
+=======
+>>>>>>> parent of a723e0a... Including commits frmo merges #29
 ### GitHub [#28](https://github.com/tomasbjerre/git-changelog-lib/issues/28) Crash when git repo has no master branch
 
 **Finding first commit in repo as parents of HEAD**
 
  * Was looking at parents of master, which may not exist. 
 
-[87fe4044e587b1b](https://github.com/tomasbjerre/git-changelog-lib/commit/87fe4044e587b1b) Tomas Bjerre *2016-06-02 17:55:59*
+[afe82368a25accc](https://github.com/tomasbjerre/git-changelog-lib/commit/afe82368a25accc) Tomas Bjerre *2016-06-02 17:54:26*
 
 
 ### Other changes